- add fps output - set default dataset loader n_worker = 16 - remove wogoal config - modify DH policy - new all task evaluation environment kitchen-all-v0

+++ spirl/utils/pytorch_utils.py
@@ -195,9 +195,9 @@ def ar2ten(array, device, dtype=None):
         return array
 
     if isinstance(array, np.ndarray):
-        tensor = torch.from_numpy(array).to(device)
+        tensor = torch.from_numpy(array).float().to(device)
     else:
-        tensor = torch.tensor(array).to(device)
+        tensor = torch.tensor(array).float().to(device)
     if dtype is not None:
         tensor = tensor.to(dtype)
     return tensor
